+++ src/org/nlogo/extensions/chrest/domainSpecifics/scene/GetDomainSpecificColFromSceneSpecificCol.java
@@ -42,7 +42,7 @@ public class GetDomainSpecificColFromSceneSpecificCol extends DefaultReporter {
    */
   @Override
   public Object report(Argument[] args, Context context) throws ExtensionException, LogoException {
-    return ((Scene)args[0].get()).getDomainSpecificColFromSceneSpecificCol(args[1].getIntValue());
+    return (double)((Scene)args[0].get()).getDomainSpecificColFromSceneSpecificCol(args[1].getIntValue());
   }
   
 }
