+++ src/org/nlogo/extensions/chrest/domainSpecifics/scene/GetDomainSpecificRowFromSceneSpecificRow.java
@@ -42,7 +42,7 @@ public class GetDomainSpecificRowFromSceneSpecificRow extends DefaultReporter{
    */
   @Override
   public Object report(Argument[] args, Context context) throws ExtensionException, LogoException {
-    return ((Scene)args[0].get()).getDomainSpecificRowFromSceneSpecificRow(args[1].getIntValue());
+    return (double)((Scene)args[0].get()).getDomainSpecificRowFromSceneSpecificRow(args[1].getIntValue());
   }
   
 }
