+++ src/org/nlogo/extensions/chrest/domainSpecifics/scene/GetSceneSpecificRowFromDomainSpecificRow.java
@@ -42,6 +42,6 @@ public class GetSceneSpecificRowFromDomainSpecificRow extends DefaultReporter{
    */
   @Override
   public Object report(Argument[] args, Context context) throws ExtensionException, LogoException {
-    return ((Scene)args[0].get()).getSceneSpecificRowFromDomainSpecificRow(args[1].getIntValue());
+    return (double)((Scene)args[0].get()).getSceneSpecificRowFromDomainSpecificRow(args[1].getIntValue());
   }
 }
